chore: relocate styles from EventById to types folder
